Enhance user registration with improved logging, validation, and repository configuration
