+++ applications/app-service/src/main/resources/application.yaml
@@ -49,3 +49,12 @@ springdoc:
       description: "This is the API for Crediya Auth Microservice"
   swagger-ui:
     path: /swagger-ui.html
+
+
+logging:
+  level:
+    org.springframework.r2dbc.core.DefaultDatabaseClient: DEBUG
+    org.springframework.transaction.interceptor: DEBUG
+    org.springframework.transaction.reactive: DEBUG
+    org.springframework.transaction: DEBUG
+    io.r2dbc.h2: DEBUG
